Added userId verfication in all routes: Users will only be able to manipulate their own data and cannot interfere with other users data

+++ backend/index.js
@@ -55,12 +55,12 @@ app.post("/create/lunch",authenticate, async (req, res)=>{
      await LunchRequest.create({
         name: createPayload.name,
         college: createPayload.college,
-        phoneNumber: createPayload.phoneNumber
+        phoneNumber: createPayload.phoneNumber,
+        userId: req.user.userId
      })
      res.json({
         msg: "Lunch Request created successfully"
      })
-
 })
 
 app.post("/create/housekeeping",authenticate,async (req, res)=>{
@@ -78,6 +78,7 @@ app.post("/create/housekeeping",authenticate,async (req, res)=>{
      await HousekeepingRequest.create({
         name: createPayload.name,
         roomNumber: createPayload.roomNumber,
+        userId: req.user.userId
      })
      res.json({
         msg: "HouseKeeping Request created successfully"
@@ -101,6 +102,7 @@ app.post("/create/complain",authenticate,async (req, res)=>{
         name: createPayload.name,
         issue: createPayload.issue,
         roomNumber: createPayload.roomNumber,
+        userId: req.user.userId
      })
      res.json({
         msg: "Complain registered successfully"
@@ -112,7 +114,8 @@ app.post("/create/complain",authenticate,async (req, res)=>{
 
 app.get("/requests/lunch",authenticate, async (req, res)=>{
     //route to get all the lunch requests
-    const lunchReq = await LunchRequest.find();
+    let userId = req.user.userId;
+    const lunchReq = await LunchRequest.find({userId});
     res.json({
         lunchReq
     })
@@ -120,7 +123,8 @@ app.get("/requests/lunch",authenticate, async (req, res)=>{
 
 app.get("/requests/housekeeping",authenticate,async (req, res)=>{
     //route to get all the housekeeping requests
-    const housekeepingReq = await HousekeepingRequest.find();
+    let userId = req.user.userId;
+    const housekeepingReq = await HousekeepingRequest.find({userId});
     res.json({
         housekeepingReq
     })
@@ -129,7 +133,8 @@ app.get("/requests/housekeeping",authenticate,async (req, res)=>{
 
 app.get("/requests/complains",authenticate, async (req, res)=>{
     //route to get all the complains filed by you
-    const complainReq = await ComplainRequest.find();
+    let userId = req.user.userId;
+    const complainReq = await ComplainRequest.find({userId});
     res.json({
         complainReq
     })
@@ -140,7 +145,8 @@ app.get("/requests/complains",authenticate, async (req, res)=>{
 
 app.put("/update/lunch/:id",authenticate, async (req, res)=>{
     //route to update the lunch requests
-    const {id}= req.params;
+    const {id}= req.params; //this is the object id from mongo
+    let userId = req.user.userId;//userid unique to every user
     const updatePayload = req.body;
     const parsePayload = updateLunchReq.safeParse(updatePayload);
 
@@ -150,6 +156,12 @@ app.put("/update/lunch/:id",authenticate, async (req, res)=>{
     }
 
     try{
+        const lunchRequest = await LunchRequest.findById(id);
+        // Check if the lunch request belongs to the user making the request
+        if (lunchRequest.userId.toString() !== userId) {
+            // Assuming lunchRequest.userId is an ObjectId, convert it to string to compare
+            return res.status(403).json({ msg: "User not authorized to update this lunch request" });
+        }
         const updatedLunch = await LunchRequest.findByIdAndUpdate(id, updatePayload, {new : true });
 
         if(!updatedLunch){
@@ -169,6 +181,7 @@ app.put("/update/lunch/:id",authenticate, async (req, res)=>{
 app.put("/update/housekeeping/:id",authenticate, async (req, res)=>{
     //route to update the housekeeping requests
     const {id} = req.params;
+    let userId = req.user.userId;//userid unique to every user
     const updatePayload = req.body;
     const parsePayload = updateHousekeepingReq.safeParse(updatePayload);
 
@@ -178,6 +191,12 @@ app.put("/update/housekeeping/:id",authenticate, async (req, res)=>{
         })
     }
     try{
+        const housekeepingRequest = await HousekeepingRequest.findById(id);
+        // Check if the lunch request belongs to the user making the request
+        if (housekeepingRequest.userId.toString() !== userId) {
+            // Assuming lunchRequest.userId is an ObjectId, convert it to string to compare
+            return res.status(403).json({ msg: "User not authorized to update this lunch request" });
+        }
         const updatedHousekeeping = await HousekeepingRequest.findByIdAndUpdate(id, updatePayload, {new: true});
 
         if(!updatedHousekeeping){
@@ -187,10 +206,11 @@ app.put("/update/housekeeping/:id",authenticate, async (req, res)=>{
         }
 
         res.json({
-            msg: "HouseKeeping request successfully added",
+            msg: "HouseKeeping request successfully updated",
             updatedHousekeeping
         })
     }catch(err){
+        console.error(`Error updating housekeeping request: ${err.message}`);
         res.status(500).json({msg: "Server error"})
     }
 })
@@ -198,6 +218,8 @@ app.put("/update/housekeeping/:id",authenticate, async (req, res)=>{
 app.put("/update/complains/:id",authenticate,async (req, res)=>{
     //route to update the complains 
     const {id} = req.params;
+    let userId = req.user.userId;//userid unique to every user
+
     const updatePayload = req.body;
     const parsePayload = updateComplainReq.safeParse(updatePayload);
 
@@ -207,6 +229,12 @@ app.put("/update/complains/:id",authenticate,async (req, res)=>{
     }
 
     try {
+        const complainRequest = await ComplainRequest.findById(id);
+        // Check if the lunch request belongs to the user making the request
+        if (complainRequest.userId.toString() !== userId) {
+            // Assuming lunchRequest.userId is an ObjectId, convert it to string to compare
+            return res.status(403).json({ msg: "User not authorized to update this lunch request" });
+        }
         const updatedComplain = await ComplainRequest.findByIdAndUpdate(id, updatePayload, {new: true});
 
         if(!updatedComplain){
@@ -228,8 +256,16 @@ app.put("/update/complains/:id",authenticate,async (req, res)=>{
 
 app.delete("/delete/lunch/:id",authenticate, async (req, res)=>{
     //route to delete the lunch requests
+    let userId = req.user.userId;
     const {id} = req.params;
     try {
+        const lunchRequest = await LunchRequest.findById(id);
+
+        if(lunchRequest.userId.toString() !== userId){
+            // Assuming lunchRequest.userId is an ObjectId, convert it to string to compare
+            return res.status(403).json({msg:"User not authorized to delete this lunch request"})
+        }
+
         await LunchRequest.findByIdAndDelete(id);
 
         res.status(200).json({msg:"Lunch request deleted successfully"})
@@ -242,10 +278,17 @@ app.delete("/delete/lunch/:id",authenticate, async (req, res)=>{
 
 app.delete("/delete/housekeeping/:id",authenticate, async (req, res)=>{
     //route to delete the housekeeping requests
+    let userId = req.user.userId;
     const {id} = req.params;
     try {
+        const housekeepingReq = await HousekeepingRequest.findById(id);
+
+        if(housekeepingReq.userId.toString() !== userId){
+            return res.status(403).json({msg:"User not authorized to delete this housekeeping request"})
+        }
         await HousekeepingRequest.findByIdAndDelete(id);
         res.status(200).json({msg:"Housekeeping request successfully deleted"})
+    
     } catch (error) {
         res.status(500).json({msg:"Error deleting housekeeping request"})
     }
@@ -254,8 +297,15 @@ app.delete("/delete/housekeeping/:id",authenticate, async (req, res)=>{
 
 app.delete("/delete/complains/:id",authenticate,async (req, res)=>{
     //route to delete the complains requests
+    let userId = req.user.userId;
     const {id} = req.params;
     try {
+        const complainReq = await ComplainRequest.findById(id);
+
+        if(complainReq.userId.toString() !== userId){
+            return res.status(403).json({msg:"User not authorized to delete this complain request"})
+        }
+
         await ComplainRequest.findByIdAndDelete(id);
         res.status(200).json({msg:"Complain request successfully deleted"})
     } catch (error) {
